Update the 147 mirs to current mirbase ids.
Only 144 (of the 147 mirs) have a current mirbase id, which is correct
because:
The only original mirbase ids that could not be mapped to a current
mirbase id were mir-276, mir-2811 and mir-2812.  mir-2811 is a dead id,
mir-2812 is an id for bro-mir-2812 (i.e. wrong species), and mir-276
legitimately looked like it should have been searched for under
dme-miR-276a and dme-miR-276b.  However, since both of those mirs were
independently searched for, no loss of information occurred by excluding
mir-276.